Refactor ARTIS density tests to use fixture for file path and add legacy reading test
