+++ tardis/io/model/readers/tests/test_artis_readers.py
@@ -1,3 +1,4 @@
+import numpy as np
 import numpy.testing as npt
 from astropy import units as u
 
@@ -8,12 +9,15 @@ from tardis.io.model.readers.artis import (
 )
 
 
-def test_artis_density_reader(example_model_file_dir):
+@pytest.fixture
+def artis_density_fname(example_model_file_dir):
+    return example_model_file_dir / "artis_model.dat"
+
+
+def test_artis_density_reader(artis_density_fname: str):
     # Using a test ARTIS density file.
     # File: tardis_artis_density_test.dat
-    time_model, velocity, mean_density = read_artis_density(
-        example_model_file_dir / "artis_model.dat"
-    )
+    time_model, velocity, mean_density = read_artis_density(artis_density_fname)
     # Check that time is recognized as time
     assert time_model.unit.physical_type == "time"
     # Check velocity unit is cm/s
@@ -57,4 +61,16 @@ def test_artis_model_reader(example_model_file_dir):
     )
 
 
-# ...existing code if any...
+def test_simple_legacy_read_artis_density(artis_density_fname: str):
+    time_of_model, velocity, mean_density, isotope_mass_fractions = read_artis_density(
+        artis_density_fname, legacy_return=True
+    )
+
+    assert np.isclose(0.00114661 * u.day, time_of_model, atol=1e-7 * u.day)
+    assert np.isclose(
+        mean_density[23],
+        0.2250048 * u.g / u.cm**3,
+        atol=1.0e-6 * u.g / u.cm**3,
+    )
+    assert len(mean_density) == 69
+    assert len(velocity) == len(mean_density) + 1
